+++ rotkehlchen/assets/exchanges_mappings/iconomi.py
@@ -3,8 +3,6 @@ from rotkehlchen.constants.resolver import evm_address_to_identifier, strethaddr
 from rotkehlchen.types import ChainID, EvmTokenKind
 
 WORLD_TO_ICONOMI = COMMON_ASSETS_MAPPINGS | {
-    # In Rotkehlchen LUNA-2 is Terra Luna but in Iconomi it's LUNA
-    'LUNA-2': 'LUNA',
     # make sure iconomi matches ADX latest contract
     strethaddress_to_identifier('0xADE00C28244d5CE17D72E40330B1c318cD12B7c3'): 'ADX',
     # make sure iconomi matces ANT latest contract
@@ -23,7 +21,6 @@ WORLD_TO_ICONOMI = COMMON_ASSETS_MAPPINGS | {
     strethaddress_to_identifier('0x88dF592F8eb5D7Bd38bFeF7dEb0fBc02cf3778a0'): 'TRB',
     # EDG is Edgeless in iconomi
     strethaddress_to_identifier('0x08711D3B02C8758F2FB3ab4e80228418a7F8e39c'): 'EDG',
-    'SOL-2': 'SOL',
     strethaddress_to_identifier('0xF970b8E36e23F7fC3FD752EeA86f8Be8D83375A6'): 'RCN',
     'ONE-2': 'ONE',
     strethaddress_to_identifier('0x3432B6A60D23Ca0dFCa7761B7ab56459D9C964D0'): 'FXS',